Core/Maps - Improvements to Cinematic function  - Moves to own class for reading, storage and player subclass  - Proper destruction handling for player (ensure cinematic is ended)  - Timeout for cinematic if it reaches past the end without completing  - boost::filesystem::path used for path/filename transform  - Correct for assert trigger under certain circumstances  - Other changes previously suggested
