+++ src/server/game/DataStores/DBCEnums.h
@@ -19,6 +19,22 @@
 #ifndef DBCENUMS_H
 #define DBCENUMS_H
 
+#pragma pack(push, 1)
+
+struct DBCPosition2D
+{
+    float X;
+    float Y;
+};
+
+struct DBCPosition3D
+{
+    float X;
+    float Y;
+    float Z;
+};
+
+#pragma pack(pop)
 enum LevelLimit
 {
     // Client expected level limitation, like as used in DBC item max levels for "until max player level"
